requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.5
django-appconf==1.0.4
django-ckeditor==5.9.0
django-comments-dab==1.6.1
django-comments-xtd==2.6.1
django-contrib-comments==1.9.2
django-crispy-forms==1.9.0
django-croppie==1.3
django-extensions==3.0.2
django-filebrowser==3.13.1
django-grappelli==2.14.2
django-hitcount==1.3.2
django-image-cropping==1.4.0
django-js-asset==1.2.2
django-likes==2.0.1
django-markdown-deux==1.0.5
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.16
draftjs-exporter==2.1.7
easy-thumbnails==2.7
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
l18n==2018.5
Markdown==3.2.2
markdown2==2.3.9
oauthlib==3.1.0
Pillow==7.1.2
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.3
XlsxWriter==1.2.8
zipp==3.1.0
